file rename and extra file checks to remove err msgs

+++ moveSrcs.sh
@@ -1,2 +1,9 @@
-mv ~/.zshrc ~/.zshrc.old
-mv ~/.vimrc ~/.vimrc.old
+if [[ -f ${HOME}/.zshrc ]]
+then 
+    mv ~/.zshrc ~/.zshrc.old
+fi
+if [[ -f ${HOME}/.vimrc ]]
+then 
+    mv ~/.vimrc ~/.vimrc.old
+fi
+
